Fix local add to dash
